+++ .playroom/frame.tsx
@@ -1,6 +1,3 @@
-import '../packages/overdrive/lib/reset/globalFonts.css';
-import '../packages/overdrive/lib/reset';
-
 import * as React from 'react';
 import { StrictMode } from 'react';
 import { OverdriveProvider } from '@autoguru/overdrive';