+++ packages/overdrive/lib/index.ts
@@ -1,6 +1,6 @@
-export * from './components';
+export * from './components/index';
 
-export { useMedia } from './hooks/useMedia';
-export { useResponsiveValue } from './hooks/useResponsiveValue';
-export { useAttachedBoxes } from './hooks/useAttachedBoxes';
-export { useId, arrayRingLookup } from './utils';
+export { useMedia } from './hooks/useMedia/useMedia';
+export { useResponsiveValue } from './hooks/useResponsiveValue/useResponsiveValue';
+export { useAttachedBoxes } from './hooks/useAttachedBoxes/useAttachedBoxes';
+export { useId, arrayRingLookup } from './utils/index';