+++ packages/overdrive/lib/reset/index.ts
@@ -1,1 +1,0 @@
-import './globalReset.css';