+++ packages/overdrive/lib/themes/tokens.ts
@@ -1,3 +1,4 @@
+import { Tokens as VanillaTokens } from '@vanilla-extract/css/dist/declarations/src/types';
 type SpaceScale = '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | 'none';
 type TextSizeScale = '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9';
 type BorderWidthScale = '1' | '2' | '3' | 'none';
@@ -14,7 +15,6 @@ interface ColourIntensityMap extends VanillaTokens {
 }
 
 export type ColourMap = Record<string, ColourValue>;
-import { Tokens as VanillaTokens } from '@vanilla-extract/css/dist/declarations/src/types';
 
 type ColourReds =
 	| 'red900'